fix: update import paths after folder reorganization
