+++ src/components/input/index.ts
@@ -1,7 +1,7 @@
 // AdvancedInput has been moved to src/features/prism/PrismInput
 // For backward compatibility, re-export from new location
-export { default as AdvancedInput, PrismInput } from '@src/features/prism/PrismInput';
-export type { PrismInputRef as AdvancedInputRef, PrismInputRef } from '@src/features/prism/PrismInput';
+export { default as AdvancedInput, PrismInput } from '@src/features/prism/conversation/PrismInput';
+export type { PrismInputRef as AdvancedInputRef, PrismInputRef } from '@src/features/prism/conversation/PrismInput';
 
 // Re-export types that are used by the input component
 export type { ToolsPromptInterface } from '@src/types';
